+++ toolium/pageelements/page_element.py
@@ -62,7 +62,9 @@ class PageElement(CommonObject):
         try:
             self._find_web_element()
         except NoSuchElementException as exception:
-            msg = "Page element of type '{}' with locator {} not found".format(type(self).__name__, self.locator)
+            parent_msg = " and parent locator '{}'".format(self.parent) if self.parent else ''
+            msg = "Page element of type '{}' with locator {}{} not found".format(type(self).__name__, self.locator,
+                                                                                 parent_msg)
             self.logger.error(msg)
             exception.msg += "\n  {}".format(msg)
             raise exception
@@ -92,10 +94,11 @@ class PageElement(CommonObject):
         :returns: page element instance
         """
         try:
-            self._web_element = self.utils.wait_until_element_visible(self.locator, timeout)
+            self.utils.wait_until_element_visible(self, timeout)
         except TimeoutException as exception:
-            msg = "Page element of type '{}' with locator {} not found or is not visible after {} seconds".format(
-                type(self).__name__, self.locator, timeout)
+            parent_msg = " and parent locator '{}'".format(self.parent, timeout) if self.parent else ''
+            msg = "Page element of type '{}' with locator {}{} not found or is not visible after {} seconds".format(
+                type(self).__name__, self.locator, parent_msg, timeout)
             self.logger.error(msg)
             exception.msg += "\n  {}".format(msg)
             raise exception
@@ -108,8 +111,7 @@ class PageElement(CommonObject):
         :returns: page element instance
         """
         try:
-            web_element = self.utils.wait_until_element_not_visible(self.locator, timeout)
-            self._web_element = web_element if web_element else None
+            self.utils.wait_until_element_not_visible(self, timeout)
         except TimeoutException as exception:
             msg = "Page element of type '{}' with locator {} is still visible after {} seconds".format(
                 type(self).__name__, self.locator, timeout)